Move api specific suites into new modules
